Divided structures into separate source files (issue #142, conversions)
